Created common base class called "Portfolio" that both Profiles and CRDs use.
