+++ arch/crd_family/MockCRDFamily.yaml
@@ -1,15 +1,6 @@
 MockCRDFamily:
   name: MockCRDFamily
   long_name: Mock CRD Family Long Name
-  revision_history:
-  - version: "0.1"
-    date: 2024-10-04
-    changes:
-    - Created to test CRDs
-  - version: "0.2"
-    date: 2024-10-05
-    changes:
-    - Also created to test CRDs
 
   introduction: |
     Here's the Mock CRD Family's introduction.
@@ -19,3 +10,5 @@ MockCRDFamily:
 
   mandatory_priv_modes:
   - M
+
+  description: Here's the Mock CRD Family's description.
