+++ backends/portfolio_doc/tasks.rb
@@ -32,7 +32,6 @@ Dir.glob("#{$root}/arch/crd/*.yaml") do |f|
     # switch to the generated CRD arch def
     arch_def = crd.to_arch_def
     crd = arch_def.crd(crd_name)
-    crd_family = crd.family
 
     version = File.basename(t.name, '.adoc').split('-')[1..].join('-')
