+++ backends/portfolio_doc/templates/family_intro.erb
@@ -1,0 +1,11 @@
+== <%= portfolio_family.name %> Introduction
+
+<%= portfolio_family.introduction %>
+
+=== <%= portfolio_family.name %> Naming Scheme
+
+<%= portfolio_family.naming_scheme %>
+
+=== <%= portfolio_family.name %> Family Description
+
+<%= portfolio_family.description %>
